test(sidebar): move sidebar locators to a dedicated file
