+++ packages/bruno-app/src/components/Sidebar/Collections/Collection/CollectionItem/DeleteCollectionItem/index.js
@@ -45,6 +45,7 @@ const DeleteCollectionItem = ({ onClose, item, collectionUid }) => {
         confirmButtonColor="danger"
         handleConfirm={onConfirm}
         handleCancel={onClose}
+        dataTestId="delete-collection-item-modal"
       >
         Are you sure you want to delete <span className="font-medium">{item.name}</span> ?
       </Modal>
