+++ tests/utils/page/index.ts
@@ -2,6 +2,7 @@ export * from './actions';
 export * from './file-mode';
 export * from './runner';
 export * from './locators';
+export * from './sidebar';
 export * from './mounting';
 export * from './preferences';
 export * from './ai';
